+++ src/js/components/internal/slider-transitioner.js
@@ -1,5 +1,5 @@
 import {translate} from '../../mixin/internal/slideshow-animations';
-import {children, clamp, createEvent, css, Deferred, includes, index, isRtl, noop, offset, position, sortBy, Transition, trigger} from 'uikit-util';
+import {children, clamp, createEvent, css, Deferred, dimensions, includes, index, isRtl, noop, position, sortBy, Transition, trigger} from 'uikit-util';
 
 export default function (prev, next, dir, {center, easing, list}) {
 
@@ -7,10 +7,10 @@ export default function (prev, next, dir, {center, easing, list}) {
 
     const from = prev
         ? getLeft(prev, list, center)
-        : getLeft(next, list, center) + offset(next).width * dir;
+        : getLeft(next, list, center) + dimensions(next).width * dir;
     const to = next
         ? getLeft(next, list, center)
-        : from + offset(prev).width * dir * (isRtl ? -1 : 1);
+        : from + dimensions(prev).width * dir * (isRtl ? -1 : 1);
 
     return {
 
@@ -36,10 +36,6 @@ export default function (prev, next, dir, {center, easing, list}) {
 
         },
 
-        stop() {
-            return Transition.stop(list);
-        },
-
         cancel() {
             Transition.cancel(list);
         },
@@ -60,7 +56,7 @@ export default function (prev, next, dir, {center, easing, list}) {
             css(list, 'transform', translate(clamp(
                 -to + (distance - distance * percent),
                 -getWidth(list),
-                offset(list).width
+                dimensions(list).width
             ) * (isRtl ? -1 : 1), 'px'));
 
             this.updateTranslates();
@@ -92,9 +88,9 @@ export default function (prev, next, dir, {center, easing, list}) {
         },
 
         getActives() {
-            return [next].concat(children(list).filter((slide, i) => {
+            return [prev || next].concat(children(list).filter(slide => {
                 const slideLeft = getElLeft(slide, list);
-                return slideLeft > from && slideLeft + offset(slide).width <= offset(list).width + from;
+                return slideLeft > from && slideLeft + dimensions(slide).width <= dimensions(list).width + from;
             }));
         },
 
@@ -127,23 +123,19 @@ function getLeft(el, list, center) {
 }
 
 export function getMax(list) {
-    return Math.max(0, getWidth(list) - offset(list).width);
+    return Math.max(0, getWidth(list) - dimensions(list).width);
 }
 
 export function getWidth(list) {
-    return children(list).reduce((right, el) => offset(el).width + right, 0);
-}
-
-export function getMaxWidth(list) {
-    return children(list).reduce((right, el) => Math.max(right, offset(el).width), 0);
+    return children(list).reduce((right, el) => dimensions(el).width + right, 0);
 }
 
 function centerEl(el, list) {
-    return offset(list).width / 2 - offset(el).width / 2;
+    return dimensions(list).width / 2 - dimensions(el).width / 2;
 }
 
 export function getElLeft(el, list) {
-    return (position(el).left + (isRtl ? offset(el).width - offset(list).width : 0)) * (isRtl ? -1 : 1);
+    return el && (position(el).left + (isRtl ? dimensions(el).width - dimensions(list).width : 0)) * (isRtl ? -1 : 1) || 0;
 }
 
 function triggerUpdate(el, type, data) {