+++ src/js/components/slider-parallax.js
@@ -1,5 +1,5 @@
 import Parallax from '../mixin/parallax';
-import {css, endsWith, noop, query, Transition} from 'uikit-util';
+import {css, endsWith, fastdom, noop, query, Transition} from 'uikit-util';
 
 export default {
 
@@ -20,22 +20,6 @@ export default {
     events: [
 
         {
-
-            name: 'itemshown',
-
-            self: true,
-
-            el() {
-                return this.item;
-            },
-
-            handler() {
-                css(this.$el, this.getCss(.5));
-            }
-
-        },
-
-        {
             name: 'itemin itemout',
 
             self: true,
@@ -46,15 +30,14 @@ export default {
 
             handler({type, detail: {percent, duration, timing, dir}}) {
 
-                Transition.cancel(this.$el);
-                css(this.$el, this.getCss(getCurrent(type, dir, percent)));
-
-                Transition.start(this.$el, this.getCss(isIn(type)
-                    ? .5
-                    : dir > 0
-                        ? 1
-                        : 0
-                ), duration, timing).catch(noop);
+                fastdom.read(() => {
+                    const propsFrom = this.getCss(getCurrentPercent(type, dir, percent));
+                    const propsTo = this.getCss(isIn(type) ? .5 : dir > 0 ? 1 : 0);
+                    fastdom.write(() => {
+                        css(this.$el, propsFrom);
+                        Transition.start(this.$el, propsTo, duration, timing).catch(noop);
+                    });
+                });
 
             }
         },
@@ -84,8 +67,10 @@ export default {
             },
 
             handler({type, detail: {percent, dir}}) {
-                Transition.cancel(this.$el);
-                css(this.$el, this.getCss(getCurrent(type, dir, percent)));
+                fastdom.read(() => {
+                    const props = this.getCss(getCurrentPercent(type, dir, percent));
+                    fastdom.write(() => css(this.$el, props));
+                });
             }
         }
 
@@ -97,7 +82,7 @@ function isIn(type) {
     return endsWith(type, 'in');
 }
 
-function getCurrent(type, dir, percent) {
+function getCurrentPercent(type, dir, percent) {
 
     percent /= 2;